backend/requirements.txt
# requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
mistralai
reportlab==4.1.0
pandas==2.2.0
openpyxl==3.1.2